% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myBoot2.R
\name{myboot2}
\alias{myboot2}
\title{myBoot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Number of bootstrap iterations}

\item{x}{The sample you want to bootstrap from}

\item{fun}{The function or parameter of interest}

\item{alpha}{Determines size of interval 0.05 = 95 percent confidence interval}

\item{cx}{Determines size of text}

\item{...}{Allows for more argument calls such as xlab, ylab, etc}
}
\value{
lower and upper bounds of interval, function of interest, and original sample invisibly
}
\description{
myBoot2
}
\examples{
sam = c(1,1,1,2,2,2,2,3,3,3,4,4)
statistic = myboot2(x = sam, fun = "median")
}
