% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLTGrapher.R
\name{myclt}
\alias{myclt}
\title{Central Limit Theorem Plotter}
\usage{
myclt(n, iter)
}
\arguments{
\item{n}{number of samples used}

\item{iter}{number of times we run the function for the samples}
}
\value{
Histogram of r uniform distribution showing concepts of the central limit theorem
}
\description{
Central Limit Theorem Plotter
}
\examples{
myclt(n=10,iter=10000)
}
