% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{mymlnorm}
\alias{mymlnorm}
\title{mymlnorm}
\usage{
mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{Sample of vector data points}

\item{mu}{Vector of potential mean values}

\item{sig}{Vector of potential standard deviation values}

\item{...}{Allows for additional arguments to be passed to the function}
}
\value{
A list with x, coord, and maxl
}
\description{
mymlnorm
}
\examples{
mymlnorm(x=c(10,12,13,15,12,11,10),mu=seq(10,15,length=1000),sig=seq(0.1,4,length=1000),lwd=2,labcex=1)
}
