% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{Mean}

\item{sigma}{Standard Deviation}

\item{a}{Right endpoint for area}
}
\value{
A list containing the mean, standard deviation, and standard deviation.
}
\description{
myncurve
}
\examples{
myncurve(5, 10, 2)
}
