% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuadPlotter.R
\name{myplot}
\alias{myplot}
\title{Quadratic Function Plot}
\usage{
myplot(x)
}
\arguments{
\item{x}{Real Number}
}
\value{
Quadratic Equation
}
\description{
Quadratic Function Plot
}
\examples{
myplot(x = 1)
}
