% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myquadpt2.R
\name{myquadpt2}
\alias{myquadpt2}
\title{Simple Quad + 2}
\usage{
myquadpt2(x)
}
\arguments{
\item{x}{vector of quantitative numbers}
}
\value{
vector
}
\description{
Simple Quad + 2
}
\examples{
myquadpt2(1:10)
}
