% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Number of seats on the flight.}

\item{gamma}{probability of a plane being overbooked.}

\item{p}{probability of someone who bought a ticket showing up.}
}
\value{
A named list including n (number of tickets to be sold) derived discretely and normaly, N, p, and gamma. Also makes a graph for the discretely and normally distributed values.
}
\description{
ntickets
}
\examples{
ntickets(N=400, gamma=0.02, p=0.95)
}
