% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleDistributionPlot.R
\name{SampleDistPlot}
\alias{SampleDistPlot}
\title{Sample Distribution Plot}
\usage{
SampleDistPlot(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{Sample size for each iteration}

\item{iter}{Number of iterations}

\item{time}{the pause between iterations}
}
\value{
Nothing, intended purpose is for visualization
}
\description{
Sample Distribution Plot
}
\examples{
mySample30 = SampleDistPlot(n=1000, iter=30, time = 1)
}
